requirements.txt
python-dotenv
openai
requests
wikipedia
langchain
langchain-community
langchain-openai
langchainhub
tavily-python